Final Build

+++ Data Paws/Assets/Scripts/StartupManager.cs
@@ -1,14 +1,17 @@
 using UnityEngine;
 
-// Deletes all saved player preferences when the game starts
 public class StartupManager : MonoBehaviour
 {
-    void Start()
-    {
-        // Clear all stored PlayerPrefs data
-        PlayerPrefs.DeleteAll();
+    private static bool prefsResetThisSession = false;
 
-        // Force PlayerPrefs to write changes to disk
-        PlayerPrefs.Save();
+    void Awake()
+    {
+        if (!prefsResetThisSession)
+        {
+            PlayerPrefs.DeleteAll();
+            PlayerPrefs.Save();
+
+            prefsResetThisSession = true;
+        }
     }
 }
